requirements.txt
dnspython>=2.1
DTLSSocket>=0.1.12
PyYAML>=5.4.1
